pyproject.toml
[project]
name = "1-2-find-video-source"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.12.0",
    "numpy>=2.3.4",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.2",
    "youtube-transcript-api>=1.2.3",
]
